+++ CHANGELOG.md
@@ -9,6 +9,12 @@ project will use [Semantic Versioning](https://semver.org) for non-beta releases
 changes may be made on minor version bumps, and feature additions / bug-fixes
 will be made on patch versions.
 
+## [Unreleased]
+
+### Added
+- Added `lisp_list_append()` and `lisp_list_set_{left|right}` for mutating and
+  constructing lists more easily.
+
 ## [1.0.0] - 2018-08-16
 
 First 1.0 release. The API should now be stable enough for you to use!